# swagger.yaml
openapi: 3.0.3
info:
  title: PDF to PowerPoint Converter API
  description: |
    A FastAPI application for creating PDF files and converting them to PowerPoint presentations using MVC architecture.
    
    ## Features
    - **PDF Creation**: Generate PDF files from text content with custom styling
    - **PDF to PowerPoint Conversion**: Convert PDF files to PowerPoint presentations
    - **File Management**: Upload, download, list, and delete files
    - **RESTful API**: Clean and intuitive API endpoints
    
    ## Authentication
    This API currently does not require authentication for demonstration purposes.
    In production, consider implementing JWT tokens or API keys.
    
    ## Rate Limiting
    No rate limiting is currently implemented. Consider adding rate limiting for production use.
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.example.com
    description: Production server
...
paths:
...
  /convert/pdf-to-pptx:
    post:
      tags:
        - Conversion Operations
      summary: Convert PDF to PowerPoint
      description: Convert a PDF file to PowerPoint presentation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConversionRequest'
            examples:
              basic:
                summary: Basic Conversion
                value:
                  pdf_path: "outputs/pdfs/sample_document.pdf"
                  output_name: "converted_presentation"
                  include_images: true
              custom_name:
                summary: Custom Output Name
                value:
                  pdf_path: "outputs/pdfs/report.pdf"
                  output_name: "my_presentation"
                  include_images: false
      responses:
        '200':
          description: PDF converted to PowerPoint successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversionResponse'
        '400':
          description: Bad request - PDF file not found or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /convert/list-pptx:
    get:
      tags:
        - Conversion Operations
      summary: List PowerPoint Files
      description: Get a list of all available PowerPoint files
      responses:
        '200':
          description: PowerPoint files retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  pptx_files:
                    type: array
                    items:
                      $ref: '#/components/schemas/PPTXFileInfo'
                  count:
                    type: integer
                    example: 3
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /convert/download-pptx/{filename}:
    get:
      tags:
        - Conversion Operations
      summary: Download PowerPoint
      description: Download a PowerPoint file by filename
      parameters:
        - name: filename
          in: path
          required: true
          description: Name of the PowerPoint file to download
          schema:
            type: string
            example: "converted_presentation.pptx"
      responses:
        '200':
          description: PowerPoint file downloaded successfully
          content:
            application/vnd.openxmlformats-officedocument.presentationml.presentation:
              schema:
                type: string
                format: binary
        '404':
          description: PowerPoint file not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /convert/delete-pptx/{filename}:
    delete:
      tags:
        - Conversion Operations
      summary: Delete PowerPoint
      description: Delete a PowerPoint file by filename
      parameters:
        - name: filename
          in: path
          required: true
          description: Name of the PowerPoint file to delete
          schema:
            type: string
            example: "converted_presentation.pptx"
      responses:
        '200':
          description: PowerPoint file deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "PowerPoint file converted_presentation.pptx deleted successfully"
        '404':
          description: PowerPoint file not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /convert/create-sample-pptx:
    post:
      tags:
        - Conversion Operations
      summary: Create Sample PowerPoint
      description: Create a sample PowerPoint presentation for testing
      responses:
        '200':
          description: Sample PowerPoint created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Sample PowerPoint created: Sample_Presentation_20250815_125408.pptx"
                  file_path:
                    type: string
                    example: "outputs/pptx/Sample_Presentation_20250815_125408.pptx"
                  filename:
                    type: string
                    example: "Sample_Presentation_20250815_125408.pptx"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  "/cleanup/start":
    post:
      tags:
        - File Cleanup
      summary: Start Cleanup Service
      description: Start the automatic cleanup service that deletes files after 2 hours
      responses:
        '200':
          description: Cleanup service started successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Cleanup service started. Files will be deleted after 2 hours."
                  cleanup_interval_hours:
                    type: integer
                    example: 2
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  "/cleanup/stop":
    post:
      tags:
        - File Cleanup
      summary: Stop Cleanup Service
      description: Stop the automatic cleanup service
      responses:
        '200':
          description: Cleanup service stopped successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Cleanup service stopped."
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  "/cleanup/status":
    get:
      tags:
        - File Cleanup
      summary: Get Cleanup Status
      description: Get the current status of the cleanup service and file statistics
      responses:
        '200':
          description: Cleanup status retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  status:
                    type: object
                    properties:
                      service_running:
                        type: boolean
                        example: true
                      cleanup_interval_hours:
                        type: integer
                        example: 2
                      current_time:
                        type: string
                        format: date-time
                        example: "2025-08-15T13:00:00.000000"
                      cutoff_time:
                        type: string
                        format: date-time
                        example: "2025-08-15T11:00:00.000000"
                      files_to_be_deleted:
                        type: object
                        properties:
                          pdf_files:
                            type: integer
                            example: 2
                          pptx_files:
                            type: integer
                            example: 1
                          total:
                            type: integer
                            example: 3
                      total_files:
                        type: object
                        properties:
                          pdf_files:
                            type: integer
                            example: 5
                          pptx_files:
                            type: integer
                            example: 3
                          total:
                            type: integer
                            example: 8
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  "/cleanup/manual":
    post:
      tags:
        - File Cleanup
      summary: Manual Cleanup
      description: Perform manual cleanup of files older than 2 hours
      responses:
        '200':
          description: Manual cleanup completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Manual cleanup completed. Deleted 3 files."
                  result:
                    type: object
                    properties:
                      deleted_files:
                        type: array
                        items:
                          type: object
                          properties:
                            filename:
                              type: string
                              example: "Sample_Document_20250815_110000.pdf"
                            type:
                              type: string
                              example: "PDF"
                            size_bytes:
                              type: integer
                              example: 1853
                            deleted_at:
                              type: string
                              format: date-time
                              example: "2025-08-15T13:00:00.000000"
                      total_deleted:
                        type: integer
                        example: 3
                      errors:
                        type: array
                        items:
                          type: string
                      cleanup_time:
                        type: string
                        format: date-time
                        example: "2025-08-15T13:00:00.000000"
                      cutoff_time:
                        type: string
                        format: date-time
                        example: "2025-08-15T11:00:00.000000"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  "/cleanup/config":
    get:
      tags:
        - File Cleanup
      summary: Get Cleanup Configuration
      description: Get the current cleanup service configuration
      responses:
        '200':
          description: Cleanup configuration retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  config:
                    type: object
                    properties:
                      pdf_directory:
                        type: string
                        example: "outputs/pdfs"
                      pptx_directory:
                        type: string
                        example: "outputs/pptx"
                      cleanup_interval_hours:
                        type: integer
                        example: 2
                      service_running:
                        type: boolean
                        example: true
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    ConversionRequest:
      type: object
      required:
        - pdf_path
      properties:
        pdf_path:
          type: string
          description: Path to the PDF file to convert
          example: "outputs/pdfs/sample_document.pdf"
        output_name:
          type: string
          description: Name for the output PowerPoint file
          example: "converted_presentation"
          maxLength: 100
        include_images:
          type: boolean
          description: Whether to include images in conversion
          example: true
          default: true
    ConversionResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "PowerPoint created successfully: converted_presentation.pptx"
        pptx_path:
          type: string
          example: "outputs/pptx/converted_presentation.pptx"
        converted_at:
          type: string
          format: date-time
          example: "2025-08-15T12:54:00.191290"
...
    PPTXFileInfo:
      type: object
      properties:
        filename:
          type: string
          example: "converted_presentation.pptx"
        size_bytes:
          type: integer
          example: 30955
        created_at:
          type: number
          format: double
          example: 1755242640.1912327
        modified_at:
          type: number
          format: double
          example: 1755242640.1912327
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "PDF file not found"
        message:
          type: string
          example: "The requested PDF file could not be found"
        timestamp:
          type: string
          format: date-time
          example: "2025-08-15T12:54:42.002577"
        path:
          type: string
          example: "/pdf/download/nonexistent.pdf"
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (not currently required)
security:
  - ApiKeyAuth: []
...